refactor(storages): merge storages/constants to storages/table-meta
