+++ src/query/service/src/interpreters/interpreter_table_show_create.rs
@@ -19,13 +19,13 @@ use common_datavalues::prelude::*;
 use common_exception::Result;
 use common_planner::plans::ShowCreateTablePlan;
 use common_sql::executor::PhysicalScalar;
+use common_storages_table_meta::table::is_internal_opt_key;
 use tracing::debug;
 
 use crate::interpreters::Interpreter;
 use crate::pipelines::PipelineBuildResult;
 use crate::sessions::QueryContext;
 use crate::sessions::TableContext;
-use crate::sql::is_internal_opt_key;
 
 pub struct ShowCreateTableInterpreter {
     ctx: Arc<QueryContext>,
