+++ src/query/service/tests/it/storages/fuse/pruning.rs
@@ -28,13 +28,13 @@ use common_sql::executor::ExpressionOp;
 use common_storages_fuse::FuseTable;
 use common_storages_table_meta::meta::BlockMeta;
 use common_storages_table_meta::meta::TableSnapshot;
+use common_storages_table_meta::table::OPT_KEY_DATABASE_ID;
+use common_storages_table_meta::table::OPT_KEY_SNAPSHOT_LOCATION;
 use databend_query::interpreters::CreateTableInterpreterV2;
 use databend_query::interpreters::Interpreter;
 use databend_query::sessions::QueryContext;
 use databend_query::sessions::TableContext;
 use databend_query::sql::plans::create_table_v2::CreateTablePlanV2;
-use databend_query::sql::OPT_KEY_DATABASE_ID;
-use databend_query::sql::OPT_KEY_SNAPSHOT_LOCATION;
 use databend_query::storages::fuse::io::MetaReaders;
 use databend_query::storages::fuse::pruning::BlockPruner;
 use databend_query::storages::fuse::FUSE_OPT_KEY_BLOCK_PER_SEGMENT;
