+++ src/query/service/tests/it/storages/fuse/table.rs
@@ -23,6 +23,8 @@ use common_planner::plans::DropTableClusterKeyPlan;
 use common_planner::ReadDataSourcePlan;
 use common_planner::SourceInfo;
 use common_sql::executor::table_read_plan::ToReadDataSourcePlan;
+use common_storages_table_meta::table::OPT_KEY_DATABASE_ID;
+use common_storages_table_meta::table::OPT_KEY_SNAPSHOT_LOCATION;
 use databend_query::interpreters::AlterTableClusterKeyInterpreter;
 use databend_query::interpreters::CreateTableInterpreterV2;
 use databend_query::interpreters::DropTableClusterKeyInterpreter;
@@ -31,8 +33,6 @@ use databend_query::interpreters::InterpreterFactory;
 use databend_query::sessions::TableContext;
 use databend_query::sql::plans::create_table_v2::CreateTablePlanV2;
 use databend_query::sql::Planner;
-use databend_query::sql::OPT_KEY_DATABASE_ID;
-use databend_query::sql::OPT_KEY_SNAPSHOT_LOCATION;
 use databend_query::storages::fuse::io::MetaReaders;
 use databend_query::storages::fuse::FuseTable;
 use databend_query::stream::DataBlockStream;
