+++ src/query/service/tests/it/storages/fuse/table_test_fixture.rs
@@ -28,6 +28,7 @@ use common_sql::executor::table_read_plan::ToReadDataSourcePlan;
 use common_storage::StorageFsConfig;
 use common_storage::StorageParams;
 use common_storages_fuse::FUSE_TBL_XOR_BLOOM_INDEX_PREFIX;
+use common_storages_table_meta::table::OPT_KEY_DATABASE_ID;
 use common_streams::SendableDataBlockStream;
 use databend_query::interpreters::append2table;
 use databend_query::interpreters::CreateTableInterpreterV2;
@@ -41,7 +42,6 @@ use databend_query::sessions::QueryContext;
 use databend_query::sessions::TableContext;
 use databend_query::sql::plans::create_table_v2::CreateTablePlanV2;
 use databend_query::sql::Planner;
-use databend_query::sql::OPT_KEY_DATABASE_ID;
 use databend_query::storages::fuse::table_functions::ClusteringInformationTable;
 use databend_query::storages::fuse::table_functions::FuseSnapshotTable;
 use databend_query::storages::fuse::FUSE_TBL_BLOCK_PREFIX;
