+++ src/query/sql/Cargo.toml
@@ -39,7 +39,6 @@ common-pipeline-transforms = { path = "../pipeline/transforms" }
 common-planner = { path = "../planner" }
 common-settings = { path = "../settings" }
 common-storage = { path = "../../common/storage" }
-common-storages-constants = { path = "../storages/constants" }
 common-storages-table-meta = { path = "../storages/table-meta" }
 common-storages-view = { path = "../storages/view" }
 common-tracing = { path = "../../common/tracing" }
