+++ src/query/sql/src/lib.rs
@@ -18,5 +18,4 @@ pub mod evaluator;
 pub mod executor;
 pub mod planner;
 
-pub use common_storages_constants::*;
 pub use planner::*;
