+++ src/query/storages/fuse/fuse/Cargo.toml
@@ -30,7 +30,6 @@ common-sharing = { path = "../../../sharing" }
 common-sql = { path = "../../../../common/../query/sql" }
 common-storage = { path = "../../../../common/storage" }
 common-storages-cache = { path = "../../cache" }
-common-storages-constants = { path = "../../constants" }
 common-storages-index = { path = "../../index" }
 common-storages-table-meta = { path = "../../table-meta" }
