+++ src/query/storages/fuse/fuse/src/lib.rs
@@ -34,7 +34,6 @@ use common_catalog::table::Table;
 use common_catalog::table::TableStatistics;
 pub use common_catalog::table_context::TableContext;
 use common_catalog::table_mutator::TableMutator;
-pub use common_storages_constants::*;
 pub use constants::*;
 pub use fuse_part::ColumnLeaf;
 pub use fuse_part::ColumnLeaves;
