+++ src/query/storages/fuse/fuse/src/operations/navigate.rs
@@ -20,12 +20,12 @@ use common_exception::ErrorCode;
 use common_exception::Result;
 use common_meta_app::schema::TableStatistics;
 use common_storages_table_meta::meta::TableSnapshot;
+use common_storages_table_meta::table::OPT_KEY_SNAPSHOT_LOCATION;
 use futures::TryStreamExt;
 
 use crate::io::MetaReaders;
 use crate::io::SnapshotHistoryReader;
 use crate::FuseTable;
-use crate::OPT_KEY_SNAPSHOT_LOCATION;
 
 impl FuseTable {
     pub async fn navigate_to_time_point(
