+++ src/query/storages/table-meta/src/lib.rs
@@ -16,3 +16,4 @@
 
 pub mod caches;
 pub mod meta;
+pub mod table;
