+++ src/query/storages/table-meta/src/table/mod.rs
@@ -1,4 +1,4 @@
-//  Copyright 2022 Datafuse Labs.
+// Copyright 2021 Datafuse Labs.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,10 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#![deny(unused_crate_dependencies)]
+mod table_keys;
+mod table_prefix;
 
-mod table_option_keys;
-mod table_storage_prefix;
-
-pub use table_option_keys::*;
-pub use table_storage_prefix::*;
+pub use table_keys::*;
+pub use table_prefix::*;
